Fix wrong indentation
The twig extra markdown uses a little trick to remove indentation at the
start of each line. For this it uses first line indentation as
reference. Our indentation is not representative inside included files
so it was breaking on long indented line.

By removing all indentation, we remove that problem.